Extract local deploy embedding from workspace preparation
