+++ src/Foundry/DependencyInjection/ServiceCollectionExtensions.cs
@@ -32,6 +32,7 @@ public static class ServiceCollectionExtensions
         services.AddSingleton<WinPeToolResolver>();
         services.AddSingleton<WinPeProcessRunner>();
         services.AddSingleton<WinPeDriverPackageService>();
+        services.AddSingleton<IWinPeLocalDeployEmbeddingService, WinPeLocalDeployEmbeddingService>();
         services.AddSingleton<WinPeUsbMediaService>();
         services.AddSingleton<IWinPeWorkspacePreparationService, WinPeWorkspacePreparationService>();
         services.AddSingleton<IMediaOutputService, MediaOutputService>();
