+++ src/Foundry/Services/WinPe/IWinPeLocalDeployEmbeddingService.cs
@@ -1,0 +1,10 @@
+namespace Foundry.Services.WinPe;
+
+internal interface IWinPeLocalDeployEmbeddingService
+{
+    Task<WinPeResult> ProvisionAsync(
+        string mountedImagePath,
+        WinPeArchitecture architecture,
+        string workingDirectoryPath,
+        CancellationToken cancellationToken);
+}
