+++ src/Foundry/Services/WinPe/WinPeWorkspacePreparationService.cs
@@ -1,5 +1,4 @@
 using System.Globalization;
-using System.IO.Compression;
 using System.Text;
 using Microsoft.Extensions.Logging;
 
@@ -10,6 +9,7 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
     private readonly IWinPeDriverCatalogService _driverCatalogService;
     private readonly IWinPeDriverInjectionService _driverInjectionService;
     private readonly WinPeDriverPackageService _driverPackageService;
+    private readonly IWinPeLocalDeployEmbeddingService _localDeployEmbeddingService;
     private readonly WinPeToolResolver _toolResolver;
     private readonly WinPeProcessRunner _processRunner;
     private readonly ILogger<WinPeWorkspacePreparationService> _logger;
@@ -18,6 +18,7 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
         IWinPeDriverCatalogService driverCatalogService,
         IWinPeDriverInjectionService driverInjectionService,
         WinPeDriverPackageService driverPackageService,
+        IWinPeLocalDeployEmbeddingService localDeployEmbeddingService,
         WinPeToolResolver toolResolver,
         WinPeProcessRunner processRunner,
         ILogger<WinPeWorkspacePreparationService> logger)
@@ -25,6 +26,7 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
         _driverCatalogService = driverCatalogService;
         _driverInjectionService = driverInjectionService;
         _driverPackageService = driverPackageService;
+        _localDeployEmbeddingService = localDeployEmbeddingService;
         _toolResolver = toolResolver;
         _processRunner = processRunner;
         _logger = logger;
@@ -316,7 +318,7 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
             cancellationToken).ConfigureAwait(false);
         _logger.LogInformation("Wrote bootstrap script into mounted WinPE image. System32Path={System32Path}", system32);
 
-        WinPeResult localDeployProvisioning = await ProvisionLocalDeployArchiveInImageAsync(
+        WinPeResult localDeployProvisioning = await _localDeployEmbeddingService.ProvisionAsync(
             session.MountDirectoryPath,
             artifact.Architecture,
             artifact.WorkingDirectoryPath,
@@ -514,53 +516,6 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
                 "Required WinPE optional components were not found in ADK.",
                 $"No required component CAB was found under '{ocRoot}'.");
     }
-    private async Task<WinPeResult> ProvisionLocalDeployArchiveInImageAsync(
-        string mountedImagePath,
-        WinPeArchitecture architecture,
-        string workingDirectoryPath,
-        CancellationToken cancellationToken)
-    {
-        if (!IsEnabledEnvironmentFlag(Environment.GetEnvironmentVariable(WinPeDefaults.LocalDeployEnableEnvironmentVariable)))
-        {
-            return WinPeResult.Success();
-        }
-
-        _logger.LogInformation("Provisioning local Foundry.Deploy archive into mounted WinPE image. Architecture={Architecture}", architecture);
-        WinPeResult<string> archiveResult = await ResolveLocalDeployArchivePathAsync(architecture, workingDirectoryPath, cancellationToken).ConfigureAwait(false);
-        if (!archiveResult.IsSuccess)
-        {
-            _logger.LogWarning("Failed to resolve local Foundry.Deploy archive path. Code={ErrorCode}, Message={ErrorMessage}",
-                archiveResult.Error?.Code,
-                archiveResult.Error?.Message);
-            return WinPeResult.Failure(archiveResult.Error!);
-        }
-
-        string destinationPath = Path.Combine(mountedImagePath, WinPeDefaults.EmbeddedDeployArchivePathInImage);
-        string? destinationDirectory = Path.GetDirectoryName(destinationPath);
-        if (string.IsNullOrWhiteSpace(destinationDirectory))
-        {
-            return WinPeResult.Failure(
-                WinPeErrorCodes.InternalError,
-                "Failed to resolve destination path for local Foundry.Deploy archive.",
-                $"Destination: '{destinationPath}'.");
-        }
-
-        try
-        {
-            Directory.CreateDirectory(destinationDirectory);
-            File.Copy(archiveResult.Value!, destinationPath, overwrite: true);
-            return WinPeResult.Success();
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to copy local Foundry.Deploy archive into mounted WinPE image. DestinationPath={DestinationPath}", destinationPath);
-            return WinPeResult.Failure(
-                WinPeErrorCodes.BuildFailed,
-                "Failed to copy local Foundry.Deploy archive into mounted WinPE image.",
-                ex.ToString());
-        }
-    }
-
     private WinPeResult ProvisionBundledSevenZipInImage(
         string mountedImagePath,
         WinPeArchitecture architecture)
@@ -678,207 +633,6 @@ internal sealed class WinPeWorkspacePreparationService : IWinPeWorkspacePreparat
                 WinPeErrorCodes.BuildFailed,
                 "Failed to provision Foundry.Deploy expert configuration into mounted WinPE image.",
                 ex.ToString());
-        }
-    }
-
-    private async Task<WinPeResult<string>> ResolveLocalDeployArchivePathAsync(
-        WinPeArchitecture architecture,
-        string workingDirectoryPath,
-        CancellationToken cancellationToken)
-    {
-        string configuredArchivePath = (Environment.GetEnvironmentVariable(WinPeDefaults.LocalDeployArchiveEnvironmentVariable) ?? string.Empty).Trim();
-        if (!string.IsNullOrWhiteSpace(configuredArchivePath))
-        {
-            if (!File.Exists(configuredArchivePath))
-            {
-                return WinPeResult<string>.Failure(
-                    WinPeErrorCodes.ValidationFailed,
-                    "Configured local Foundry.Deploy archive was not found.",
-                    $"Set {WinPeDefaults.LocalDeployArchiveEnvironmentVariable} to an existing .zip file. Path: '{configuredArchivePath}'.");
-            }
-
-            return WinPeResult<string>.Success(configuredArchivePath);
-        }
-
-        string configuredProjectPath = (Environment.GetEnvironmentVariable(WinPeDefaults.LocalDeployProjectEnvironmentVariable) ?? string.Empty).Trim();
-        string projectPath;
-        if (!string.IsNullOrWhiteSpace(configuredProjectPath))
-        {
-            if (!File.Exists(configuredProjectPath))
-            {
-                return WinPeResult<string>.Failure(
-                    WinPeErrorCodes.ValidationFailed,
-                    "Configured Foundry.Deploy project file was not found.",
-                    $"Set {WinPeDefaults.LocalDeployProjectEnvironmentVariable} to an existing .csproj file. Path: '{configuredProjectPath}'.");
-            }
-
-            projectPath = configuredProjectPath;
-        }
-        else if (!TryFindFoundryDeployProjectPath(out projectPath))
-        {
-            return WinPeResult<string>.Failure(
-                WinPeErrorCodes.ValidationFailed,
-                "Unable to locate Foundry.Deploy project for local WinPE embedding.",
-                $"Set {WinPeDefaults.LocalDeployArchiveEnvironmentVariable} to a .zip archive or {WinPeDefaults.LocalDeployProjectEnvironmentVariable} to Foundry.Deploy.csproj.");
-        }
-
-        string runtimeIdentifier = architecture.ToDotnetRuntimeIdentifier();
-        string localWorkspace = Path.Combine(workingDirectoryPath, "FoundryDeployLocal");
-        string publishDirectory = Path.Combine(localWorkspace, "publish", runtimeIdentifier);
-        string archivePath = Path.Combine(localWorkspace, $"Foundry.Deploy-{runtimeIdentifier}.zip");
-
-        TryDeleteDirectory(publishDirectory);
-        Directory.CreateDirectory(publishDirectory);
-
-        try
-        {
-            if (File.Exists(archivePath))
-            {
-                File.Delete(archivePath);
-            }
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to prepare local workspace for Foundry.Deploy archive generation. ArchivePath={ArchivePath}", archivePath);
-            return WinPeResult<string>.Failure(
-                WinPeErrorCodes.BuildFailed,
-                "Failed to prepare local workspace for Foundry.Deploy archive generation.",
-                ex.ToString());
-        }
-
-        string publishArgs = string.Join(" ",
-            "publish",
-            WinPeProcessRunner.Quote(projectPath),
-            "-c", "Release",
-            "-r", runtimeIdentifier,
-            "--self-contained", "true",
-            "/p:PublishSingleFile=true",
-            "/p:EnableCompressionInSingleFile=true",
-            "/p:IncludeNativeLibrariesForSelfExtract=true",
-            "/p:IncludeAllContentForSelfExtract=true",
-            "/p:DebugType=None",
-            "/p:GenerateDocumentationFile=false",
-            "-o", WinPeProcessRunner.Quote(publishDirectory));
-
-        WinPeProcessExecution publish = await _processRunner.RunAsync(
-            "dotnet",
-            publishArgs,
-            workingDirectoryPath,
-            cancellationToken).ConfigureAwait(false);
-
-        if (!publish.IsSuccess)
-        {
-            return WinPeResult<string>.Failure(
-                WinPeErrorCodes.BuildFailed,
-                "Failed to publish Foundry.Deploy for local WinPE embedding.",
-                publish.ToDiagnosticText());
-        }
-
-        string executablePath = Path.Combine(publishDirectory, "Foundry.Deploy.exe");
-        if (!File.Exists(executablePath))
-        {
-            return WinPeResult<string>.Failure(
-                WinPeErrorCodes.BuildFailed,
-                "Foundry.Deploy publish output is incomplete.",
-                $"Expected executable: '{executablePath}'.");
-        }
-
-        try
-        {
-            ZipFile.CreateFromDirectory(publishDirectory, archivePath, CompressionLevel.Optimal, includeBaseDirectory: false);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Failed to create Foundry.Deploy archive from publish output. PublishDirectory={PublishDirectory}, ArchivePath={ArchivePath}", publishDirectory, archivePath);
-            return WinPeResult<string>.Failure(
-                WinPeErrorCodes.BuildFailed,
-                "Failed to create Foundry.Deploy archive from local publish output.",
-                ex.ToString());
-        }
-
-        return WinPeResult<string>.Success(archivePath);
-    }
-
-    private static bool TryFindFoundryDeployProjectPath(out string projectPath)
-    {
-        foreach (string root in GetProjectDiscoveryRoots())
-        {
-            if (TryResolveFoundryDeployProjectPath(root, out projectPath))
-            {
-                return true;
-            }
-        }
-
-        projectPath = string.Empty;
-        return false;
-    }
-
-    private static IEnumerable<string> GetProjectDiscoveryRoots()
-    {
-        string current = Directory.GetCurrentDirectory();
-        if (!string.IsNullOrWhiteSpace(current))
-        {
-            yield return current;
-        }
-
-        string baseDirectory = AppContext.BaseDirectory;
-        if (!string.IsNullOrWhiteSpace(baseDirectory) && !baseDirectory.Equals(current, StringComparison.OrdinalIgnoreCase))
-        {
-            yield return baseDirectory;
-        }
-    }
-
-    private static bool TryResolveFoundryDeployProjectPath(string startDirectory, out string projectPath)
-    {
-        var current = new DirectoryInfo(startDirectory);
-        while (current is not null)
-        {
-            string candidate = Path.Combine(current.FullName, "src", "Foundry.Deploy", "Foundry.Deploy.csproj");
-            if (File.Exists(candidate))
-            {
-                projectPath = candidate;
-                return true;
-            }
-
-            current = current.Parent;
-        }
-
-        projectPath = string.Empty;
-        return false;
-    }
-
-    private static bool IsEnabledEnvironmentFlag(string? value)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-        {
-            return false;
-        }
-
-        return value.Trim() switch
-        {
-            "1" => true,
-            "true" => true,
-            "TRUE" => true,
-            "yes" => true,
-            "YES" => true,
-            "on" => true,
-            "ON" => true,
-            _ => false
-        };
-    }
-
-    private static void TryDeleteDirectory(string path)
-    {
-        try
-        {
-            if (Directory.Exists(path))
-            {
-                Directory.Delete(path, true);
-            }
-        }
-        catch
-        {
-            // Best-effort cleanup.
         }
     }
